Remove concurrently tools
Although concurrently support cross platform it's not necessary for this
project. Simple `&` should be enough to run tasks in parallel.
NOTE: The downsize of using `&` is that dev server has to be manually killed after task `npm run e2e` is completed.
The web server process remains in the background.